refactor:add interface for better service class decoupling
