+++ src/main/java/br/com/biblioteca/services/LivroServiceImpl.java
@@ -1,0 +1,53 @@
+package br.com.biblioteca.services;
+
+import br.com.biblioteca.dtos.LivroRequestDTO;
+import br.com.biblioteca.dtos.LivroResponseDTO;
+import br.com.biblioteca.models.LivroModel;
+import br.com.biblioteca.repositories.LivroRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+import java.util.stream.Collectors;
+
+@Service
+public class LivroServiceImpl implements LivroService {
+
+    private final LivroRepository livroRepository;
+
+    @Autowired
+    public LivroServiceImpl(LivroRepository livroRepository) {
+        this.livroRepository = livroRepository;
+    }
+    @Override
+    public LivroResponseDTO cadastraLivro(LivroRequestDTO livroDTO) {
+        LivroModel livro = new LivroModel(null, livroDTO.getTitulo(), livroDTO.getAutor(), livroDTO.getAnoPublicacao());
+        LivroModel livroSalvo = livroRepository.save(livro);
+        return new LivroResponseDTO(livroSalvo.getId(), livroSalvo.getTitulo(), livroSalvo.getAutor(), livroSalvo.getAnoPublicacao());
+    }
+    @Override
+    public LivroResponseDTO atualizarLivros(Long id, LivroRequestDTO livroDTO) {
+        LivroModel livro = livroRepository.findById(id)
+                .orElseThrow(() -> new RuntimeException("Livro não encontrado"));
+
+        livro.setTitulo(livroDTO.getTitulo());
+        livro.setAutor(livroDTO.getAutor());
+        livro.setAnoPublicacao(livroDTO.getAnoPublicacao());
+
+        LivroModel livroAtualizado = livroRepository.save(livro);
+        return new LivroResponseDTO(livroAtualizado.getId(), livroAtualizado.getTitulo(), livroAtualizado.getAutor(), livroAtualizado.getAnoPublicacao());
+    }
+    @Override
+    public List<LivroResponseDTO> listarLivros() {
+        return livroRepository.findAll()
+                .stream()
+                .map(livro -> new LivroResponseDTO(livro.getId(), livro.getTitulo(), livro.getAutor(), livro.getAnoPublicacao()))
+                .collect(Collectors.toList());
+    }
+    @Override
+    public void deletar(Long id) {
+        livroRepository.deleteById(id);
+    }
+
+
+}
